+++ metrics.ts
@@ -1,6 +1,7 @@
 import Site from "./site.ts";
 import { Page } from "./filesystem.ts";
 import { brightGreen, gray } from "./deps/colors.ts";
+import { join } from "./deps/path.ts";
 
 /**
  * Class to collect and return performance metrics
@@ -78,6 +79,7 @@ export default class Metrics {
    * Save the metrics data in a file
    */
   async save(file: string) {
-    await Deno.writeTextFile(file, JSON.stringify(this.entries));
+    const path = join(this.site.options.cwd, file);
+    await Deno.writeTextFile(path, JSON.stringify(this.entries));
   }
 }